+++ apps/block_scout_web/assets/js/pages/address.js
@@ -8,6 +8,8 @@ import { createStore, connectElements } from '../lib/redux_helpers.js'
 import { updateAllCalculatedUsdValues } from '../lib/currency.js'
 import { loadTokenBalanceDropdown } from '../lib/token_balance_dropdown'
 import '../lib/token_balance_dropdown_search'
+import '../lib/async_listing_load'
+import '../app'
 
 export const initialState = {
   channelDisconnected: false,