+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [#2988](https://github.com/poanetwork/blockscout/pull/2988) - Fix web manifest accessibility
 - [#2967](https://github.com/poanetwork/blockscout/pull/2967) - Fix styles loading for firefox
 - [#2950](https://github.com/poanetwork/blockscout/pull/2950) - Add `creationMethod` to `EthereumJSONRPC.Parity.Trace.Action.entry_to_elixir`
+- [#2944](https://github.com/poanetwork/blockscout/pull/2944) - Split js logic into multiple files
 - [#2897](https://github.com/poanetwork/blockscout/pull/2897) - remove duplicate indexes
 - [#2883](https://github.com/poanetwork/blockscout/pull/2883) - Fix long contracts names